themes:...:rekor-logo: add alpha channel.
I borrowed this logo for a diagram and to my dismay there's no alpha
channel on it (preposterous!). Add an alpha channel-version of the same
image in case we ever change background color to not-white.

I'm sure an actual designer would probably do a better job, but for now
this seems reasonable to me...

Signed-off-by: Santiago Torres <santiago@archlinux.org>